Use consistent usize index type for all child iterators (#527)
* Use consistent usize index type for all child iterators

BooleanOpChildIter, FloatOpChildIter, and StringOpChildIter used u8 for
their index field, while BitVecOpChildIter already used usize. Using u8
limits n-ary operations (like And/Or) to 255 children and is inconsistent.

Change the index field to usize in all three iterators and remove the
now-unnecessary `as usize` casts.


* Fix rustfmt formatting in BooleanOpChildIter

Collapse single-expression match arms to single lines per rustfmt style.


---------

+++ crates/clarirs_core/src/ast/float.rs
@@ -32,7 +32,7 @@ pub type FloatAst<'c> = AstRef<'c, FloatOp<'c>>;
 
 pub struct FloatOpChildIter<'a, 'c> {
     op: &'a FloatOp<'c>,
-    index: u8,
+    index: usize,
 }
 
 impl<'c> FloatOp<'c> {
@@ -114,7 +114,7 @@ impl<'a, 'c> ExactSizeIterator for FloatOpChildIter<'a, 'c> {
 
             FloatOp::FpFP(..) | FloatOp::ITE(..) => 3,
         };
-        total.saturating_sub(self.index as usize)
+        total.saturating_sub(self.index)
     }
 }
